fix start app
