+++ ewm_main/src/main/java/ru/practicum/main_server/service/impl/CompilationServiceImpl.java
@@ -1,6 +1,7 @@
 package ru.practicum.main_server.service.impl;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
@@ -26,6 +27,8 @@ import static java.util.Objects.isNull;
 public class CompilationServiceImpl implements CompilationService {
 
     private final CompilationRepository compilationRepository;
+
+    @Autowired
     private final CheckEntity check;
 
     @Override
