+++ ewm_stats/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 server.port=9090
 
 logging.level.org.springframework.web.client.RestTemplate=info
-#logging.level.org.apache.http=DEBUG
+logging.level.org.apache.http=DEBUG
 logging.level.httpclient.wire=info
 
 spring.jpa.hibernate.ddl-auto=none
